// packages/web/src/generated/AgentAnswerView.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AgentAnswerPart } from "./AgentAnswerPart";
import type { AgentAnswerSource } from "./AgentAnswerSource";

export type AgentAnswerView = { parts: Array<AgentAnswerPart>, sources: Array<AgentAnswerSource>, };